changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11755 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/Views/TextFinder/GoToLine.h
@@ -34,8 +34,8 @@
 - (IBAction) orderFrontGoToPanel:(id)sender;
 - (IBAction) jumpButtonClicked:(id)sender;
 
-- (void) showLine:(unsigned)lineNumber;
-- (BOOL) showCharacter:(unsigned)charIndex granularity:(NSSelectionGranularity)granularity;
+- (void) showLine:(NSUInteger)lineNumber;
+- (BOOL) showCharacter:(NSUInteger)charIndex granularity:(NSSelectionGranularity)granularity;
 
 
 @end
